integrate changes from evm module
