+++ hedera-node/hedera-mono-service/src/test/java/com/hedera/node/app/service/mono/store/contracts/precompile/ERC721PrecompilesTest.java
@@ -272,6 +272,9 @@ class ERC721PrecompilesTest {
                 .when(() -> EntityIdUtils.tokenIdFromEvmAddress(nonFungibleTokenAddr.toArray()))
                 .thenReturn(token);
         entityIdUtils
+                .when(() -> EntityIdUtils.tokenIdFromEvmAddress(nonFungibleTokenAddr))
+                .thenReturn(token);
+        entityIdUtils
                 .when(
                         () ->
                                 EntityIdUtils.contractIdFromEvmAddress(
